# openAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Excel GPT Middleware API",
    "version": "1.1.0",
    "description": "Express API middleware for Microsoft Graph Excel with name-based resolution, recursive search, Excel engine formatting, intelligent find/replace, and rename operations.\nNotes:\n- All Excel and rename endpoints are mounted under /api/excel. Health endpoints are under /health.\n- Name-vs-ID: All endpoints accept human-friendly names (driveName, folderName, fileName, sheetName) with automatic name-to-ID resolution and backward compatibility for IDs.\n- Path disambiguation: When duplicate file names exist, you can pass itemPath (e.g., '/Folder/Subfolder/My.xlsx').\n- Error envelope is consistent: { status: \"success\" | \"error\" | \"partial_success\" | \"preview\" | \"multiple_matches\" | \"no_matches\", ... }. Success responses include status: \"success\" and may include data."
  },
  "servers": [
    {
      "url": "https://excel-middleware.vercel.app",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/excel/worksheets": {
      "get": {
        "summary": "List worksheets in a workbook",
        "operationId": "getWorksheets",
        "parameters": [
          { "in": "query", "name": "siteId", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint Site ID override" },
          { "in": "query", "name": "siteUrl", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint Site URL override" },
          { "in": "query", "name": "sharepointHostname", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint hostname override (tenant.sharepoint.com)" },
          { "in": "query", "name": "sharepointSiteName", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint site name override (e.g., Finance)" },
          { "in": "query", "name": "driveName", "schema": { "type": "string" } },
          { "in": "query", "name": "itemName", "schema": { "type": "string" }, "description": "Workbook file name (e.g., Report.xlsx)." },
          { "in": "query", "name": "itemPath", "schema": { "type": "string" }, "description": "Full path to disambiguate duplicates (e.g., /Folder/Sub/Report.xlsx)." }
        ],
        "responses": {
          "200": { "description": "List success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListSheetsSuccess" } } } },
          "404": { "description": "Drive or file not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
...
    "/api/excel/read-table": {
      "post": {
        "summary": "Read a table by name or range",
        "operationId": "readTable",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReadTableRequest" } } } },
        "responses": {
          "200": { "description": "Read table success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReadTableSuccess" } } } },
          "400": { "description": "Bad request / ambiguity", "content": { "application/json": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ErrorWithAvailableSheets" }, { "$ref": "#/components/schemas/ErrorWithAvailableDrives" } ] } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
    "/api/excel/add-table-rows": {
      "post": {
        "summary": "Append rows to a table",
        "operationId": "addTableRows",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddTableRowsRequest" } } } },
        "responses": {
          "200": { "description": "Append success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeSuccessWithData" } } } },
          "400": { "description": "Bad request / ambiguity", "content": { "application/json": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ErrorWithAvailableSheets" }, { "$ref": "#/components/schemas/ErrorWithAvailableDrives" } ] } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
...
    "/api/excel/find-replace": {
      "post": {
        "summary": "Intelligent find & replace with optional preview and highlighting",
        "operationId": "findReplace",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FindReplaceRequest" } } } },
        "responses": {
          "200": { "description": "Operation success (may include change log)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FindReplaceSuccess" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "404": { "description": "Targets not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "409": { "description": "Multiple matches requiring user disambiguation", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MultipleMatchError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
    "/api/excel/search-text": {
      "post": {
        "summary": "Search text without replacement",
        "operationId": "searchText",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SearchTextRequest" } } } },
        "responses": {
          "200": { "description": "Search results", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SearchTextSuccess" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
    "/api/excel/analyze-scope": {
      "get": {
        "summary": "Analyze workbook structure for find/replace planning",
        "operationId": "analyzeScope",
        "parameters": [
          { "in": "query", "name": "siteId", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint Site ID override" },
          { "in": "query", "name": "siteUrl", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint Site URL override" },
          { "in": "query", "name": "sharepointHostname", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint hostname override (tenant.sharepoint.com)" },
          { "in": "query", "name": "sharepointSiteName", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint site name override (e.g., Finance)" },
          { "in": "query", "name": "driveName", "schema": { "type": "string" } },
          { "in": "query", "name": "itemName", "schema": { "type": "string" } },
          { "in": "query", "name": "itemPath", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Scope metadata", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeSuccessWithData" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
    "/api/excel/format": {
      "post": {
        "summary": "Apply formatting, formulas, and advanced Excel engine operations",
        "operationId": "excelFormat",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExcelFormatRequest" } } } },
        "responses": {
          "200": { "description": "Formatting applied", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeSuccessWithData" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
    "/api/excel/validate-formula": {
      "post": {
        "summary": "Validate Excel formula syntax",
        "operationId": "validateFormula",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidateFormulaRequest" } } } },
        "responses": {
          "200": { "description": "Validation result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidateFormulaSuccess" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
    "/api/excel/cell-info": {
      "get": {
        "summary": "Get comprehensive cell information",
        "operationId": "cellInfo",
        "parameters": [
          { "in": "query", "name": "siteId", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint Site ID override" },
          { "in": "query", "name": "siteUrl", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint Site URL override" },
          { "in": "query", "name": "sharepointHostname", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint hostname override (tenant.sharepoint.com)" },
          { "in": "query", "name": "sharepointSiteName", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint site name override (e.g., Finance)" },
          { "in": "query", "name": "driveName", "schema": { "type": "string" } },
          { "in": "query", "name": "itemName", "schema": { "type": "string" } },
          { "in": "query", "name": "itemPath", "schema": { "type": "string" } },
          { "in": "query", "name": "sheetName", "schema": { "type": "string" } },
          { "in": "query", "name": "cellAddress", "schema": { "type": "string" }, "required": true, "description": "A1-style address (e.g., 'B7')." }
        ],
        "responses": {
          "200": { "description": "Cell info result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeSuccessWithData" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
...
    "/api/excel/worksheet-info": {
      "get": {
        "summary": "Worksheet structure and metadata",
        "operationId": "worksheetInfo",
        "parameters": [
          { "in": "query", "name": "siteId", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint Site ID override" },
          { "in": "query", "name": "siteUrl", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint Site URL override" },
          { "in": "query", "name": "sharepointHostname", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint hostname override (tenant.sharepoint.com)" },
          { "in": "query", "name": "sharepointSiteName", "schema": { "type": "string" }, "required": false, "description": "Optional SharePoint site name override (e.g., Finance)" },
          { "in": "query", "name": "driveName", "schema": { "type": "string" } },
          { "in": "query", "name": "itemName", "schema": { "type": "string" } },
          { "in": "query", "name": "itemPath", "schema": { "type": "string" } },
          { "in": "query", "name": "sheetName", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Worksheet info", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeSuccessWithData" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
...
    "/api/excel/create-file": {
      "post": {
        "summary": "Create a new Excel file in a drive/folder",
        "operationId": "createFile",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateFileRequest" } } }
        },
        "responses": {
          "200": { "description": "File created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeSuccessWithData" } } } },
          "409": { "description": "Conflict - file exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
    "/api/excel/create-sheet": {
      "post": {
        "summary": "Add a worksheet to a workbook",
        "operationId": "createSheet",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateSheetRequest" } } }
        },
        "responses": {
          "200": { "description": "Sheet created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeSuccessWithData" } } } },
          "400": { "description": "Bad request or duplicate name", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "404": { "description": "Drive or file not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
    "/api/excel/delete-file": {
      "delete": {
        "summary": "Delete a workbook by name (use itemPath to disambiguate)",
        "operationId": "deleteFile",
        "parameters": [
          { "in": "query", "name": "siteId", "schema": { "type": "string" }, "required": false },
          { "in": "query", "name": "siteUrl", "schema": { "type": "string" }, "required": false }
        ],
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteFileRequest" } } } },
        "responses": {
          "200": { "description": "File deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeSuccessWithData" } } } },
          "409": { "description": "Multiple matches", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MultipleMatchError" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    },
    "/api/excel/delete-sheet": {
      "delete": {
        "summary": "Delete a worksheet from a workbook (not the last)",
        "operationId": "deleteSheet",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteSheetRequest" } } } },
        "responses": {
          "200": { "description": "Sheet deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeSuccessWithData" } } } },
          "400": { "description": "Cannot delete last sheet or bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnvelopeError" } } } }
        }
      }
    }
  },
  "components": {
...
    "schemas": {
...
      "EnvelopeSuccessWithData": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["success"], "example": "success" },
          "data": {}
        },
        "required": ["status"]
      },
      "EnvelopeError": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["error"], "example": "error" },
          "error": { "oneOf": [ { "type": "string" }, { "type": "object", "additionalProperties": true } ] }
        },
        "required": ["status", "error"]
      },
      "MultipleMatchError": {
        "allOf": [
          { "$ref": "#/components/schemas/EnvelopeError" },
          {
            "type": "object",
            "properties": {
              "options": { "type": "array", "items": { "type": "string" }, "description": "Candidate full paths for disambiguation" }
            }
          }
        ]
      },
      "ErrorWithAvailableSheets": {
        "allOf": [
          { "$ref": "#/components/schemas/EnvelopeError" },
          { "type": "object", "properties": { "availableSheets": { "type": "array", "items": { "type": "string" } } } }
        ]
      },
      "ErrorWithAvailableDrives": {
        "allOf": [
          { "$ref": "#/components/schemas/EnvelopeError" },
          { "type": "object", "properties": { "availableDrives": { "type": "array", "items": { "type": "string" } } } }
        ]
      },
...
      "ListSheetsSuccess": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["success"], "example": "success" },
          "sheets": { "type": "array", "items": { "type": "string" } }
        },
        "required": ["status", "sheets"]
      },
...
      "ReadTableRequest": {
        "type": "object",
        "properties": {
          "siteId": { "type": "string" },
          "siteUrl": { "type": "string" },
          "sharepointHostname": { "type": "string" },
          "sharepointSiteName": { "type": "string" },
          "driveName": { "type": "string" },
          "itemName": { "type": "string" },
          "itemPath": { "type": "string" },
          "sheetName": { "type": "string" },
          "tableName": { "type": "string" },
          "range": { "type": "string" }
        },
        "required": ["itemName"]
      },
      "ReadTableSuccess": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["success"], "example": "success" },
          "data": { "type": "object", "additionalProperties": true }
        },
        "required": ["status"]
      },
      "AddTableRowsRequest": {
        "type": "object",
        "properties": {
          "siteId": { "type": "string" },
          "siteUrl": { "type": "string" },
          "sharepointHostname": { "type": "string" },
          "sharepointSiteName": { "type": "string" },
          "driveName": { "type": "string" },
          "itemName": { "type": "string" },
          "itemPath": { "type": "string" },
          "sheetName": { "type": "string" },
          "tableName": { "type": "string" },
          "rows": { "type": "array", "items": { "type": "array", "items": {} } }
        },
        "required": ["itemName", "tableName", "rows"]
      },
...
      "FindReplaceRequest": {
        "type": "object",
        "properties": {
          "siteId": { "type": "string" },
          "siteUrl": { "type": "string" },
          "sharepointHostname": { "type": "string" },
          "sharepointSiteName": { "type": "string" },
          "driveName": { "type": "string" },
          "itemName": { "type": "string" },
          "itemPath": { "type": "string" },
          "scope": { "type": "string", "enum": ["header_only", "specific_range", "entire_sheet", "all_sheets"], "default": "entire_sheet" },
          "sheetName": { "type": "string" },
          "rangeSpec": { "type": "string" },
          "searchTerm": { "type": "string" },
          "replaceTerm": { "type": "string" },
          "confirm": { "type": "boolean", "default": false },
          "highlightChanges": { "type": "boolean", "default": false },
          "logChanges": { "type": "boolean", "default": true },
          "previewId": { "type": "string" }
        },
        "required": ["itemName", "searchTerm"]
      },
      "FindReplaceSuccess": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["success", "preview", "no_matches", "partial_success"], "example": "success" },
          "changes": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "summary": { "type": "object", "additionalProperties": true }
        },
        "required": ["status"]
      },
      "SearchTextRequest": {
        "type": "object",
        "properties": {
          "siteId": { "type": "string" },
          "siteUrl": { "type": "string" },
          "sharepointHostname": { "type": "string" },
          "sharepointSiteName": { "type": "string" },
          "driveName": { "type": "string" },
          "itemName": { "type": "string" },
          "itemPath": { "type": "string" },
          "scope": { "type": "string", "enum": ["header_only", "specific_range", "entire_sheet", "all_sheets"], "default": "entire_sheet" },
          "sheetName": { "type": "string" },
          "rangeSpec": { "type": "string" },
          "searchTerm": { "type": "string" }
        },
        "required": ["itemName", "searchTerm"]
      },
      "SearchTextSuccess": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["success", "no_matches"], "example": "success" },
          "results": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        },
        "required": ["status"]
      },
...
    }
  }
}